requirements.txt
flask
flask-cors
transformers
torch
peft
accelerate
sentencepiece
bitsandbytes
requests
websocket-client
